# lib/api-zod/src/generated/types/addCommentRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * SocialConnect API specification
 * OpenAPI spec version: 0.1.0
 */

export interface AddCommentRequest {
  /**
   * @minLength 1
   * @maxLength 500
   */
  content: string;
}