# lib/api-zod/src/generated/types/comment.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * SocialConnect API specification
 * OpenAPI spec version: 0.1.0
 */
import type { PostAuthor } from "./postAuthor";

export interface Comment {
  id: number;
  content: string;
  author: PostAuthor;
  post_id: number;
  created_at: Date;
  updated_at: Date;
}